updated close command logic
Close command now issued only if the TCP client is connected to the Dome Controller. Previously, it could enter the timeout state if the rain sensors were wet on connect, and then not issue any close commands to the dome until the timeout finished or the dome sent an opened command.